Refactor logging configs
